rm nouse res & add addTask page

+++ app/src/main/java/com/time/cat/mvp/view/TimeCatBottom.java
@@ -95,8 +95,8 @@ public class TimeCatBottom extends ViewGroup implements View.OnClickListener {
 
         setWillNotDraw(false);
 
-        mActionGap = (int) ViewUtil.dp2px(5);
-        mContentPadding = (int) ViewUtil.dp2px(10);
+        mActionGap = ViewUtil.dp2px(5);
+        mContentPadding = ViewUtil.dp2px(10);
     }
 
     private LayoutParams createLayoutParams() {
